requirements.txt
telethon==1.34.0
python-dotenv==1.0.1
requests==2.32.3
asyncio==3.4.3
